+++ tests/testthat/test-placeholder.R
@@ -1,3 +1,3 @@
 test_that("multiplication works", {
-	expect_vector(scribblr_placeholder(NULL), size = 1L)
+	expect_vector(scribblr_placeholder(NULL), ptype = character(), size = 1L)
 })